--- repository/click/kobaken/ed25519-android/1.0.0/ed25519-android-1.0.0.aar.sha1
@@ -1,1 +1,0 @@
-52a577f80a62c8c0517c52b982c7bf660d1ddb12
