--- repository/click/kobaken/ed25519-android/1.0.0/ed25519-android-1.0.0.pom
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-  <modelVersion>4.0.0</modelVersion>
-  <groupId>click.kobaken</groupId>
-  <artifactId>ed25519-android</artifactId>
-  <version>1.0.0</version>
-  <packaging>aar</packaging>
-  <dependencies>
-    <dependency>
-      <groupId>com.android.support</groupId>
-      <artifactId>appcompat-v7</artifactId>
-      <version>24.2.1</version>
-      <scope>compile</scope>
-    </dependency>
-  </dependencies>
-</project>
