--- repository/click/kobaken/ed25519-android/1.0.0/ed25519-android-1.0.0.pom.md5
@@ -1,1 +1,0 @@
-e66248b6bf78704c0fa1d6945063da18
