--- repository/click/kobaken/ed25519-android/1.0.0/ed25519-android-1.0.0.pom.sha1
@@ -1,1 +1,0 @@
-304a6646603e2e7c3ba80076f6ffb4d94fdcca50
