--- repository/click/kobaken/ed25519-android/maven-metadata.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<metadata>
-  <groupId>click.kobaken</groupId>
-  <artifactId>ed25519-android</artifactId>
-  <versioning>
-    <release>1.0.0</release>
-    <versions>
-      <version>1.0.0</version>
-    </versions>
-    <lastUpdated>20161016031816</lastUpdated>
-  </versioning>
-</metadata>
