--- repository/click/kobaken/ed25519-android/maven-metadata.xml.md5
@@ -1,1 +1,0 @@
-69d6dbb8e6c1073944c3617a2dca3253
