--- repository/click/kobaken/ed25519-android/maven-metadata.xml.sha1
@@ -1,1 +1,0 @@
-defc7e69716ce86eb45bcb47ea0974ca5aff4430
